# docs/specs/001-sakuraplayer-v1/contracts/rest-api.openapi.yaml
openapi: 3.1.0
info:
  title: SakuraPlayer API
  version: 1.2.0
  description: >-
    Private single-administrator API for the SakuraPlayer v1 backend. Video bytes
    are never proxied by this API; signed stream endpoints return 302 redirects.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /rankings:
    get:
      tags: [Discovery]
      operationId: getRanking
      description: >-
        Reads one immutable local JavDB snapshot and never contacts JavDB during
        the request. Daily, weekly, and monthly reject year. TOP250 without year
        is the overall board; explicit years are 2008 through the server current
        year. Cursors bind the selected snapshot so a current switch cannot mix pages.
      parameters:
        - name: board
          in: query
          required: true
          schema:
            type: string
            enum: [daily, weekly, monthly, top250]
        - name: year
          in: query
          description: Only valid for top250; omitted top250 means the overall board.
          schema:
            type: integer
            minimum: 2008
            maximum: 2200
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Latest successful local snapshot filtered to AVdb-backed core-ready movies.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RankingPage'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          description: No successful snapshot exists for this board/year.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RankingUnavailableError'
...
  /cache-jobs:
    get:
      tags: [Cache]
      operationId: listCacheJobs
      parameters:
        - name: status
          in: query
          schema:
            type: array
            items:
              $ref: '#/components/schemas/CacheJobStatus'
          style: form
          explode: false
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Task snapshot and fixed capacity counters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheJobPage'
...
  /settings:
    get:
      tags: [Admin]
      operationId: getSettings
      responses:
        '200':
          description: Redacted settings and connection states.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
    patch:
      tags: [Admin]
      operationId: updateSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingsPatch'
      responses:
        '200':
          description: Redacted updated settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Cursor:
      name: cursor
      in: query
      description: >-
        Opaque versioned keyset cursor bound to the endpoint's normalized query,
        filters and sort. Malformed or cross-query cursors return validation_failed.
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 24
...
  headers:
    NoStore:
      description: Authentication state and credentials must never be cached.
      required: true
      schema:
        type: string
        const: no-store
  responses:
...
    ValidationError:
      description: Request cannot be validated.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
...
    ApiError:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          pattern: '^[a-z0-9_]+$'
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        request_id:
          type: string
...
    SourceLabel:
      type: string
      enum: [subtitle, cracked, 4k, censored]
    PlaybackProgress:
      type: object
      required: [position_seconds, duration_seconds, completed, version]
      properties:
        position_seconds:
          type: number
          minimum: 0
        duration_seconds:
          type: [number, 'null']
          exclusiveMinimum: 0
        completed:
          type: boolean
        version:
          type: integer
          format: int64
          minimum: 1
          description: Authoritative server state version after the committed write.
...
    MovieSummary:
      type: object
      required: [id, number, title, labels, favorite, source_count]
      properties:
        id:
          type: string
          format: uuid
        number:
          type: string
        title:
          type: string
        title_original:
          type: [string, 'null']
        cover_url:
          type: [string, 'null']
        publish_date:
          type: [string, 'null']
          format: date
        labels:
          type: array
          maxItems: 4
          items:
            $ref: '#/components/schemas/SourceLabel'
        favorite:
          type: boolean
        source_count:
          type: integer
          minimum: 1
        progress:
          oneOf:
            - $ref: '#/components/schemas/PlaybackProgress'
            - type: 'null'
...
    RankingItem:
      type: object
      required: [rank, movie]
      properties:
        rank:
          type: integer
          minimum: 1
        movie:
          $ref: '#/components/schemas/MovieSummary'
    RankingPage:
      type: object
      required: [board, year, available_years, synced_at, items, next_cursor]
      properties:
        board:
          type: string
          enum: [daily, weekly, monthly, top250]
        year:
          type: [integer, 'null']
        available_years:
          description: Descending 2008..current year for top250; empty otherwise.
          type: array
          maxItems: 100
          items:
            type: integer
            minimum: 2008
        synced_at:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/RankingItem'
        next_cursor:
          type: [string, 'null']
    RankingUnavailableDetails:
      type: object
      required: [reason]
      additionalProperties: false
      properties:
        reason:
          type: string
          enum:
            [credentials_not_configured, credentials_invalid, never_synced, sync_failed]
        last_error_code:
          type: [string, 'null']
          pattern: '^[a-z0-9_]+$'
    RankingUnavailableError:
      allOf:
        - $ref: '#/components/schemas/ApiError'
        - type: object
          required: [details]
          properties:
            details:
              $ref: '#/components/schemas/RankingUnavailableDetails'
...
    CacheJobStatus:
      type: string
      enum:
        - queued
        - submitting
        - offlining
        - submit_uncertain
        - resolving
        - awaiting_selection
        - ready
        - cancelling
        - cleaning
        - cleanup_failed
        - failed
        - cleaned
        - detached
    RemoteMedia:
      type: object
      required: [id, candidate_id, name, size_bytes, sequence_no, is_valid]
      properties:
        id:
          type: string
          format: uuid
        candidate_id:
          type: string
          format: uuid
        name:
          type: string
        size_bytes:
          type: integer
          format: int64
        duration_seconds:
          type: [integer, 'null']
          format: int64
        sequence_no:
          type: integer
        is_valid:
          type: boolean
    SubtitleOption:
      type: object
      required: [id, media_id, name, format, selected_by_default]
      properties:
        id:
          type: string
          format: uuid
        media_id:
          type: [string, 'null']
          format: uuid
          description: Null means the subtitle is available to every media session in this manifest.
        name:
          type: string
        format:
          type: string
          enum: [srt, ass, ssa, vtt]
        language:
          type: [string, 'null']
        selected_by_default:
          type: boolean
    CacheJob:
      type: object
      required:
        [id, movie_id, source_id, status, remote_percent, media_candidates, selected_media_ids, subtitles, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        movie_id:
          type: string
          format: uuid
        source_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/CacheJobStatus'
        remote_percent:
          type: number
          minimum: 0
          maximum: 100
        error_code:
          type: [string, 'null']
        media_candidates:
          type: array
          items:
            $ref: '#/components/schemas/RemoteMedia'
        selected_media_ids:
          type: array
          items:
            type: string
            format: uuid
        subtitles:
          type: array
          items:
            $ref: '#/components/schemas/SubtitleOption'
        ready_at:
          type: [string, 'null']
          format: date-time
        expires_at:
          type: [string, 'null']
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CacheCapacity:
      type: object
      required: [running, running_limit, queued, queued_limit, ready, ready_limit]
      properties:
        running:
          type: integer
        running_limit:
          type: integer
          const: 2
        queued:
          type: integer
        queued_limit:
          type: integer
          const: 10
        ready:
          type: integer
        ready_limit:
          type: integer
          default: 20
    CacheJobPage:
      type: object
      required: [items, capacity]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CacheJob'
        capacity:
          $ref: '#/components/schemas/CacheCapacity'
        next_cursor:
          type: [string, 'null']
...
    ProviderState:
      type: object
      required: [configured, status]
      properties:
        configured:
          type: boolean
        status:
          type: string
          enum: [unknown, available, unavailable, credentials_invalid]
        last_checked_at:
          type: [string, 'null']
          format: date-time
        last_error_code:
          type: [string, 'null']
    JavdbSettings:
      allOf:
        - $ref: '#/components/schemas/ProviderState'
        - type: object
          required: [password_configured, version]
          properties:
            username:
              type: [string, 'null']
            password_configured:
              type: boolean
            version:
              type: integer
              minimum: 0
    AiSettings:
      allOf:
        - $ref: '#/components/schemas/ProviderState'
        - type: object
          required: [api_key_configured, version]
          properties:
            base_url:
              type: [string, 'null']
              format: uri
            model:
              type: [string, 'null']
            timeout_seconds:
              type: [integer, 'null']
              minimum: 1
              maximum: 600
            api_key_configured:
              type: boolean
            version:
              type: integer
              minimum: 0
    SyncRunState:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [never, running, succeeded, failed]
        release_id:
          type: [string, 'null']
        started_at:
          type: [string, 'null']
          format: date-time
        completed_at:
          type: [string, 'null']
          format: date-time
        last_successful_at:
          type: [string, 'null']
          format: date-time
        next_scheduled_at:
          type: [string, 'null']
          format: date-time
        last_error_code:
          type: [string, 'null']
    AvdbSyncStatus:
      type: object
      required: [incremental_30d, full_reconcile]
      properties:
        incremental_30d:
          $ref: '#/components/schemas/SyncRunState'
        full_reconcile:
          $ref: '#/components/schemas/SyncRunState'
    Settings:
      type: object
      required:
        [cache_ttl_hours, ready_cache_limit, metadata_concurrency, metadata_timeout_seconds, javdb, ai, providers, avdb_sync]
      properties:
        cache_ttl_hours:
          type: integer
          minimum: 1
          maximum: 168
          default: 24
        ready_cache_limit:
          type: integer
          default: 20
        metadata_concurrency:
          type: integer
          const: 3
        metadata_timeout_seconds:
          type: integer
          const: 600
        javdb:
          $ref: '#/components/schemas/JavdbSettings'
        ai:
          $ref: '#/components/schemas/AiSettings'
        providers:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ProviderState'
        avdb_sync:
          $ref: '#/components/schemas/AvdbSyncStatus'
    SettingsPatch:
      type: object
      minProperties: 1
      properties:
        cache_ttl_hours:
          type: integer
          minimum: 1
          maximum: 168
        javdb:
          $ref: '#/components/schemas/JavdbSettingsCommand'
        ai:
          $ref: '#/components/schemas/AiSettingsCommand'
      additionalProperties: false
    SettingClearCommand:
      type: object
      required: [action, expected_version]
      properties:
        action:
          const: clear
        expected_version:
          type: integer
          minimum: 1
      additionalProperties: false
    JavdbSettingsReplace:
      type: object
      required: [action, expected_version, username, password]
      properties:
        action:
          const: replace
        expected_version:
          type: integer
          minimum: 0
        username:
          type: string
          minLength: 1
          maxLength: 255
        password:
          type: string
          minLength: 1
          maxLength: 1024
      additionalProperties: false
    JavdbSettingsCommand:
      oneOf:
        - $ref: '#/components/schemas/JavdbSettingsReplace'
        - $ref: '#/components/schemas/SettingClearCommand'
    AiSettingsReplace:
      type: object
      required: [action, expected_version, base_url, api_key, model, timeout_seconds]
      properties:
        action:
          const: replace
        expected_version:
          type: integer
          minimum: 0
        base_url:
          type: string
          format: uri
          maxLength: 2048
        api_key:
          type: string
          minLength: 1
          maxLength: 8192
        model:
          type: string
          minLength: 1
          maxLength: 255
        timeout_seconds:
          type: integer
          minimum: 1
          maximum: 600
      additionalProperties: false
    AiSettingsCommand:
      oneOf:
        - $ref: '#/components/schemas/AiSettingsReplace'
        - $ref: '#/components/schemas/SettingClearCommand'
...